requirements.txt
python-telegram-bot>=20.0
SQLAlchemy>=1.4
python-dotenv>=0.21.0
